requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.12.2
inflection==0.5.1
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
platformdirs==3.10.0
postgres==4.0
psycopg2-binary==2.9.6
psycopg2-pool==1.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
virtualenv==20.24.2
